refactor: :recycle: 分离cli逻辑，单独作为一个入口，后续可以考虑把升级operator和db逻辑放入其中
